Fixed overly greedy try/except handler.
Now at least returns picture, when pricing is not found. Instead of
completely exiting loop. HTML output is still negatively affected by
this however. When pricing is not found, the index is shifted and
pictures/description/prices become out of order by 1, everytime an item
quality is not returned. Need to "pop" index when index == ''